chore: refactor nits
