+++ internal/util/test/init.go
@@ -1,6 +1,11 @@
 package test
 
-import "path/filepath"
+import (
+	"fmt"
+	"path"
+	"path/filepath"
+	"runtime"
+)
 
 var (
 	kongRBACsKustomize        = "config/rbac/"
@@ -22,4 +27,13 @@ func init() {
 	kongRBACsKustomize = filepath.Join(root, kongRBACsKustomize)
 	kongGatewayRBACsKustomize = filepath.Join(root, kongGatewayRBACsKustomize)
 	kongCRDsRBACsKustomize = filepath.Join(root, kongCRDsRBACsKustomize)
+}
+
+func getRepoRoot() (string, error) {
+	_, b, _, ok := runtime.Caller(0)
+	if !ok {
+		return "", fmt.Errorf("failed to get repo root: runtime.Caller(0) failed")
+	}
+	d := filepath.Dir(path.Join(path.Dir(b), "../../")) // Number of ../ depends on the path of this file.
+	return filepath.Abs(d)
 }
